+++ docs/examples/fork-process-model/flask-gunicorn/gunicorn.config.py
@@ -16,7 +16,7 @@ from opentelemetry import trace
 from opentelemetry.exporter.otlp.trace_exporter import OTLPSpanExporter
 from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.sdk.trace.export import BatchExportSpanProcessor
+from opentelemetry.sdk.trace.export import BatchSpanProcessor
 
 bind = "127.0.0.1:8000"
 
@@ -44,7 +44,7 @@ def post_fork(server, worker):
     trace.set_tracer_provider(TracerProvider(resource=resource))
     # This uses insecure connection for the purpose of example. Please see the
     # OTLP Exporter documentation for other options.
-    span_processor = BatchExportSpanProcessor(
+    span_processor = BatchSpanProcessor(
         OTLPSpanExporter(endpoint="localhost:4317", insecure=True)
     )
     trace.get_tracer_provider().add_span_processor(span_processor)