+++ exporter/opentelemetry-exporter-otlp/tests/test_otlp_trace_exporter.py
@@ -56,7 +56,7 @@ from opentelemetry.sdk.trace import Status as SDKStatus
 from opentelemetry.sdk.trace import StatusCode as SDKStatusCode
 from opentelemetry.sdk.trace import TracerProvider, _Span
 from opentelemetry.sdk.trace.export import (
-    SimpleExportSpanProcessor,
+    SimpleSpanProcessor,
     SpanExportResult,
 )
 from opentelemetry.sdk.util.instrumentation import InstrumentationInfo
@@ -114,9 +114,7 @@ class TestOTLPSpanExporter(TestCase):
     def setUp(self):
         tracer_provider = TracerProvider()
         self.exporter = OTLPSpanExporter(insecure=True)
-        tracer_provider.add_span_processor(
-            SimpleExportSpanProcessor(self.exporter)
-        )
+        tracer_provider.add_span_processor(SimpleSpanProcessor(self.exporter))
         self.tracer = tracer_provider.get_tracer(__name__)
 
         self.server = server(ThreadPoolExecutor(max_workers=10))