+++ propagator/opentelemetry-propagator-b3/src/opentelemetry/propagators/b3/__init__.py
@@ -17,7 +17,7 @@ from re import compile as re_compile
 
 import opentelemetry.trace as trace
 from opentelemetry.context import Context
-from opentelemetry.trace.propagation.textmap import (
+from opentelemetry.propagators.textmap import (
     Getter,
     Setter,
     TextMapPropagator,
@@ -94,9 +94,9 @@ class B3Format(TextMapPropagator):
             self._trace_id_regex.fullmatch(trace_id) is None
             or self._span_id_regex.fullmatch(span_id) is None
         ):
-            ids_generator = trace.get_tracer_provider().ids_generator
-            trace_id = ids_generator.generate_trace_id()
-            span_id = ids_generator.generate_span_id()
+            id_generator = trace.get_tracer_provider().id_generator
+            trace_id = id_generator.generate_trace_id()
+            span_id = id_generator.generate_span_id()
             sampled = "0"
 
         else:
@@ -112,7 +112,7 @@ class B3Format(TextMapPropagator):
             options |= trace.TraceFlags.SAMPLED
 
         return trace.set_span_in_context(
-            trace.DefaultSpan(
+            trace.NonRecordingSpan(
                 trace.SpanContext(
                     # trace an span ids are encoded in hex, so must be converted
                     trace_id=trace_id,