+++ propagator/opentelemetry-propagator-jaeger/src/opentelemetry/propagators/jaeger/__init__.py
@@ -18,7 +18,7 @@ import urllib.parse
 import opentelemetry.trace as trace
 from opentelemetry import baggage
 from opentelemetry.context import Context, get_current
-from opentelemetry.trace.propagation.textmap import (
+from opentelemetry.propagators.textmap import (
     Getter,
     Setter,
     TextMapPropagator,
@@ -61,7 +61,7 @@ class JaegerPropagator(TextMapPropagator):
         ):
             return trace.set_span_in_context(trace.INVALID_SPAN, context)
 
-        span = trace.DefaultSpan(
+        span = trace.NonRecordingSpan(
             trace.SpanContext(
                 trace_id=int(trace_id, 16),
                 span_id=int(span_id, 16),